Fixed entering poetry-managed virtualenv
Entering the virtualenv when there is more than one poetry-managed
virtualenv was failing. I was looking at the output of `poetry env list
--full-path` but I should have instead been looking at the output of
`poetry env info` (the former lists all virtualenvs and parsing the
output was failing, it's simpler to just look at the latter since it's
only the info for the currently active virtualenv).